+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Add copy mode
+- Add mouse scroll config
 - Add quit confirmation dialog
 
 ## 0.6.0 - 2022-07-04